+++ client/src/App.css
@@ -1,4 +1,4 @@
-@import "./styles/registerLoginForm.css";
+@import "./styles/RegisterLoginForm.css";
 
 .App {
   text-align: center;
