+++ client/src/App.js
@@ -2,7 +2,8 @@ import React, { Component } from 'react';
 import logo from './logo.svg';
 import './App.css';
 import { Route } from 'react-router-dom';
-import registerLoginForm from './components/registerLoginForm';
+import RegisterLoginForm from './components/RegisterLoginForm';
+import UserList from './components/UserList';
 
 class App extends Component {
   render() {
@@ -15,9 +16,9 @@ class App extends Component {
         <p className="App-intro">
           Please implement the required code for the assignment.
         </p>
-        <Route path="/signin" component={registerLoginForm}/>
-        <Route path="/signup" component={registerLoginForm}/> 
-        {/* <Route path="/users" component={}/> */}
+        <Route path="/signin" component={RegisterLoginForm}/>
+        <Route path="/signup" component={RegisterLoginForm}/> 
+        <Route path="/users" component={UserList}/>
       </div>
     );
   }
